+++ ingestion/src/test/groovy/com/tickonomics/ingestion/filter/DeRoundingFilterSpec.groovy
@@ -5,6 +5,7 @@ import spock.lang.Specification
 import spock.lang.Subject
 
 import java.time.Instant
+import java.math.BigDecimal;
 
 class DeRoundingFilterSpec extends Specification {
 
@@ -16,11 +17,11 @@ class DeRoundingFilterSpec extends Specification {
 
     def "given tick at round mark with volume spike, when filter, then volume smoothed"() {
         given:
-            def tick = new TickData(ROUND_MARK, "SPY", 450.50, 10000, new int[0])
+            def tick = new TickData(ROUND_MARK, "SPY", BigDecimal.valueOf(450.50), 10000, new int[0])
             def recent = [
-                new TickData(ROUND_MARK.minusSeconds(5), "SPY", 450.48, 100, new int[0]),
-                new TickData(ROUND_MARK.minusSeconds(3), "SPY", 450.49, 120, new int[0]),
-                new TickData(ROUND_MARK.minusSeconds(1), "SPY", 450.50, 90, new int[0])
+                new TickData(ROUND_MARK.minusSeconds(5), "SPY", BigDecimal.valueOf(450.48), 100, new int[0]),
+                new TickData(ROUND_MARK.minusSeconds(3), "SPY", BigDecimal.valueOf(450.49), 120, new int[0]),
+                new TickData(ROUND_MARK.minusSeconds(1), "SPY", BigDecimal.valueOf(450.50), 90, new int[0])
             ]
 
         when:
@@ -33,9 +34,9 @@ class DeRoundingFilterSpec extends Specification {
 
     def "given tick not at round mark, when filter, then unchanged"() {
         given:
-            def tick = new TickData(NON_ROUND, "SPY", 450.50, 10000, new int[0])
+            def tick = new TickData(NON_ROUND, "SPY", BigDecimal.valueOf(450.50), 10000, new int[0])
             def recent = [
-                new TickData(NON_ROUND.minusSeconds(1), "SPY", 450.49, 100, new int[0])
+                new TickData(NON_ROUND.minusSeconds(1), "SPY", BigDecimal.valueOf(450.49), 100, new int[0])
             ]
 
         when:
@@ -47,10 +48,10 @@ class DeRoundingFilterSpec extends Specification {
 
     def "given tick at round mark without spike, when filter, then unchanged"() {
         given:
-            def tick = new TickData(ROUND_MARK, "SPY", 450.50, 100, new int[0])
+            def tick = new TickData(ROUND_MARK, "SPY", BigDecimal.valueOf(450.50), 100, new int[0])
             def recent = [
-                new TickData(ROUND_MARK.minusSeconds(2), "SPY", 450.49, 95, new int[0]),
-                new TickData(ROUND_MARK.minusSeconds(1), "SPY", 450.50, 105, new int[0])
+                new TickData(ROUND_MARK.minusSeconds(2), "SPY", BigDecimal.valueOf(450.49), 95, new int[0]),
+                new TickData(ROUND_MARK.minusSeconds(1), "SPY", BigDecimal.valueOf(450.50), 105, new int[0])
             ]
 
         when:
@@ -62,9 +63,9 @@ class DeRoundingFilterSpec extends Specification {
 
     def "given insufficient context ticks, when filter, then unchanged"() {
         given:
-            def tick = new TickData(ROUND_MARK, "SPY", 450.50, 10000, new int[0])
+            def tick = new TickData(ROUND_MARK, "SPY", BigDecimal.valueOf(450.50), 10000, new int[0])
             def recent = [
-                new TickData(ROUND_MARK.minusSeconds(1), "SPY", 450.50, 100, new int[0])
+                new TickData(ROUND_MARK.minusSeconds(1), "SPY", BigDecimal.valueOf(450.50), 100, new int[0])
             ]
 
         when:
